# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/boards": {
            "get": {
                "description": "Retrieves a list of all Kanban boards for the current user/context.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get all Kanban boards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/board.Board"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Kanban board with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Create Kanban Board",
                "parameters": [
                    {
                        "description": "Board entity",
                        "name": "board",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.Board"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created board entity",
                        "schema": {
                            "$ref": "#/definitions/board.Board"
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/boards/{boardId}": {
            "get": {
                "description": "Retrieves a Kanban board by its unique identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get Kanban Board by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board found",
                        "schema": {
                            "$ref": "#/definitions/board.Board"
                        }
                    },
                    "400": {
                        "description": "Invalid board ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Board not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the details of a Kanban board specified by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Update Kanban Board by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board entity to update",
                        "name": "board",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.Board"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Board not found or not updated",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the Kanban board specified by the board ID parameter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Delete a Kanban board by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Unable to parse board ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "board.Board": {
            "type": "object",
            "required": [
                "boardName",
                "description"
            ],
            "properties": {
                "boardId": {
                    "type": "integer"
                },
                "boardName": {
                    "type": "string",
                    "minLength": 3
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "minLength": 3
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "common.ErrorMsg": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}